+++ packages/integrations/vite/src/index.ts
@@ -1,3 +1,3 @@
 // Re-export from core
-export { mdxHybrid } from '@mdx-hybrid/core/vite'
-export type { VitePluginOptions } from '@mdx-hybrid/core'
+export { mdxHybrid } from '@jp-knj/mdx-hybrid-core/vite'
+export type { VitePluginOptions } from '@jp-knj/mdx-hybrid-core'