+++ benchmarks/CHANGELOG.md
@@ -1,8 +1,8 @@
-# @mdx-hybrid/benchmarks
+# @jp-knj/mdx-hybrid-benchmarks
 
 ## 0.0.1
 
 ### Patch Changes
 
 - Updated dependencies
-  - @mdx-hybrid/core@0.0.1
+  - @jp-knj/mdx-hybrid-core@0.0.1